awaf-formatter: Format immediate nested loop instrs
Add a new options parameter for processing directives so that directives
are not processed during the recursion. This is necessary to remove the
space after the opening '?' in a loop where the immediate next
instruction in the loop body is also another loop.

`[[[+++]]]` should translate to `??? awawawa!!!`. Without this options
parameter, this would have been formatted as `? ? ? awawawa!!!`.